Refactored so default type of form input is a string
