+++ src/js/interfaces/coordinatesForm.ts
@@ -1,0 +1,20 @@
+export interface SpecificDDSection {
+  rowNum: number;
+  latitude: number;
+  longitude: number;
+}
+
+export interface DDFormValues {
+  userInput: string;
+  rowNum: number;
+  coordinateType: string;
+}
+
+export interface DDSectionProps {
+  key: number;
+  ddSection: SpecificDDSection;
+  degreeSymbol: string;
+  renderRemoveButton: boolean;
+  setDDFormValues: (formValues: DDFormValues) => void;
+  setSection: (addSection: boolean) => void;
+}
